requirements.txt
django
django-environ==0.10.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-filter==23.2
drf-yasg==1.21.5